+++ src/com/hautrieu/chat/repositories/Repository.java
@@ -4,11 +4,9 @@ import java.util.function.Predicate;
 
 import com.hautrieu.chat.domains.BaseEntity;
 
-public interface Repository<T extends BaseEntity> {
-	
-	T getById(int id);
+public interface Repository<T extends BaseEntity> {	
+	T getById(long id);
     boolean add(T addingEntity);
     void deleteAll();
-
     T getFirst(Predicate<T> predicate);
 }
